requirements.txt
tensorflow
opencv-python
imutils
numpy
matplotlib
scikit-learn
customtkinter
pillow
